+++ create3search.cpp
@@ -80,7 +80,7 @@ void create3_search(create3_search_data *init_data)
                 sprintf(&salt[(i) * 2], "%02x", newSalt.b[i]);
             }
             salt[64] = 0;
-            printf("0x%s,%s,0x%s,%s_%llu\n", salt, hexAddr.c_str(), init_data->factory, g_strVersion.c_str(), init_data->total_compute / 1000 / 1000 / 1000);
+            printf("0x%s,%s,0x%s,%s_%lu\n", salt, hexAddr.c_str(), init_data->factory, g_strVersion.c_str(), (unsigned long)(init_data->total_compute / 1000 / 1000 / 1000));
 
             std::string outputDir = init_data->outputDir;
             if (outputDir == "") {
@@ -101,7 +101,7 @@ void create3_search(create3_search_data *init_data)
 				exit(1);
 			}
 
-            fprintf(out_file, "0x%s,%s,0x%s,%s_%llu", salt, hexAddr.c_str(), init_data->factory, "cuda_miner_v0.1.0", init_data->total_compute / 1000 / 1000 / 1000);
+            fprintf(out_file, "0x%s,%s,0x%s,%s_%lu", salt, hexAddr.c_str(), init_data->factory, "cuda_miner_v0.1.0", (unsigned long)(init_data->total_compute / 1000 / 1000 / 1000));
 
             fclose(out_file);
